refactor(chunkers): split diff.go into 3 focused files
